move extension to its own dir
