+++ extension/META.json.in
