+++ extension/Makefile
@@ -85,6 +85,3 @@ test-update:
 docs:
 	poetry install --no-directory --no-root
 	poetry run mkdocs serve
-
-run-worker:
-	cargo run --bin worker
