+++ extension/sql/vectorize--0.11.1--0.12.0.sql
