+++ extension/sql/vectorize--0.2.0--0.3.0.sql
