+++ extension/sql/vectorize--0.4.0--0.5.0.sql
