+++ extension/sql/vectorize--0.5.0--0.6.0.sql
