+++ extension/sql/vectorize--0.7.0--0.8.0.sql
