+++ extension/sql/vectorize--0.8.1--0.9.0.sql
