+++ extension/sql/vectorize--0.9.0--0.10.0.sql
