+++ extension/src/chat/mod.rs
