+++ extension/src/chat/ops.rs
@@ -1,7 +1,5 @@
 use crate::guc;
 use crate::search;
-use crate::types;
-use crate::types::VectorizeMeta;
 use crate::util::get_vectorize_meta_spi;
 
 use anyhow::{anyhow, Result};
@@ -12,6 +10,7 @@ use pgrx::prelude::*;
 
 use crate::chat::types::{ChatResponse, ContextualSearch, PromptTemplate, RenderedPrompt};
 use tiktoken_rs::{get_bpe_from_model, model::get_context_size, CoreBPE};
+use vectorize_core::types::{JobParams, VectorizeMeta};
 
 pub fn call_chat(
     agent_name: &str,
@@ -29,7 +28,7 @@ pub fn call_chat(
         error!("failed to get project metadata");
     };
 
-    let job_params = serde_json::from_value::<types::JobParams>(project_meta.params.clone())
+    let job_params = serde_json::from_value::<JobParams>(project_meta.params.clone())
         .unwrap_or_else(|e| error!("failed to deserialize job params: {}", e));
 
     // for various token count estimations
