+++ extension/src/chat/types.rs
