+++ extension/src/errors.rs
